requirements.txt
aiofiles==0.6.0
aiohttp==3.8.1
aiohttp-socks==0.6.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
frozenlist==1.3.0
future==0.18.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
jsonschema==3.2.0
Logbook==1.5.3
matrix-nio==0.18.7
multidict==6.0.2
pycryptodome==3.14.0
pyrsistent==0.18.1
python-socks==1.2.4
six==1.16.0
typer==0.4.0
unpaddedbase64==2.1.0
yarl==1.7.2
